+++ familyhelper-project-sdk/src/main/java/com/dwarfeng/familyhelper/project/sdk/util/Constraints.java
@@ -28,6 +28,11 @@ public final class Constraints {
      */
     public static final int LENGTH_USER = 50;
 
+    /**
+     * 简报的长度约束。
+     */
+    public static final int LENGTH_PROFILE = 20;
+
     private Constraints() {
         throw new IllegalStateException("禁止实例化");
     }